# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login a user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile"
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/foods": {
      "post": {
        "operationId": "FoodsController_create",
        "summary": "Create a new food (admin only)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Food created successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "foods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "FoodsController_findAll",
        "summary": "Get all foods with optional name filter",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter foods by name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all foods"
          }
        },
        "tags": [
          "foods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/foods/{id}": {
      "get": {
        "operationId": "FoodsController_findOne",
        "summary": "Get food by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return food by ID"
          },
          "404": {
            "description": "Food not found"
          }
        },
        "tags": [
          "foods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "FoodsController_update",
        "summary": "Update food (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Food updated successfully"
          },
          "404": {
            "description": "Food not found"
          }
        },
        "tags": [
          "foods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "FoodsController_remove",
        "summary": "Delete food (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Food deleted successfully"
          },
          "404": {
            "description": "Food not found"
          }
        },
        "tags": [
          "foods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/foods/barcode/{barcode}": {
      "get": {
        "operationId": "FoodsController_findByBarcode",
        "summary": "Get food by barcode",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return food by barcode"
          },
          "404": {
            "description": "Food not found"
          }
        },
        "tags": [
          "foods"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/ai-analysis/image": {
      "post": {
        "operationId": "AiAnalysisController_analyzeImage",
        "summary": "Analyze food from image",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "context": {
                    "type": "string",
                    "example": "This is a dinner plate with chicken and vegetables"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image analyzed successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "ai-analysis"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/ai-analysis/text": {
      "post": {
        "operationId": "AiAnalysisController_analyzeText",
        "summary": "Analyze food from text description",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeTextDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Text analyzed successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "ai-analysis"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/exercises": {
      "post": {
        "operationId": "ExercisesController_create",
        "summary": "Create a new exercise",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercise created successfully"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ExercisesController_findAll",
        "summary": "Get all exercises for user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all exercises for user"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/exercises/date": {
      "get": {
        "operationId": "ExercisesController_findByDate",
        "summary": "Get exercises for a specific date",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Date in ISO format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return exercises for date"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/exercises/calories": {
      "get": {
        "operationId": "ExercisesController_getCaloriesByDate",
        "summary": "Get total calories burned for a date",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Date in ISO format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return total calories burned for date"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/exercises/{id}": {
      "get": {
        "operationId": "ExercisesController_findOne",
        "summary": "Get exercise by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return exercise by ID"
          },
          "404": {
            "description": "Exercise not found"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ExercisesController_update",
        "summary": "Update exercise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise updated successfully"
          },
          "404": {
            "description": "Exercise not found"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ExercisesController_remove",
        "summary": "Delete exercise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise deleted successfully"
          },
          "404": {
            "description": "Exercise not found"
          }
        },
        "tags": [
          "exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Fitness Tracker API",
    "description": "API for fitness and nutrition tracking application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "Password123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "Password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Grilled Chicken Breast"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890"
          },
          "calories": {
            "type": "number",
            "example": 165
          },
          "protein": {
            "type": "number",
            "example": 31
          },
          "carbs": {
            "type": "number",
            "example": 0
          },
          "fat": {
            "type": "number",
            "example": 3.6
          },
          "servingSize": {
            "type": "number",
            "example": 100
          },
          "servingUnit": {
            "type": "string",
            "example": "g"
          }
        },
        "required": [
          "name",
          "calories",
          "protein",
          "carbs",
          "fat"
        ]
      },
      "UpdateFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Grilled Chicken Breast"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890"
          },
          "calories": {
            "type": "number",
            "example": 165
          },
          "protein": {
            "type": "number",
            "example": 31
          },
          "carbs": {
            "type": "number",
            "example": 0
          },
          "fat": {
            "type": "number",
            "example": 3.6
          },
          "servingSize": {
            "type": "number",
            "example": 100
          },
          "servingUnit": {
            "type": "string",
            "example": "g"
          }
        }
      },
...
      "AnalyzeTextDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "I had a caesar salad with grilled chicken for lunch"
          }
        },
        "required": [
          "description"
        ]
      },
      "CreateExerciseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Running"
          },
          "caloriesBurned": {
            "type": "number",
            "example": 300
          },
          "duration": {
            "type": "number",
            "example": 30
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15T17:00:00Z"
          }
        },
        "required": [
          "name",
          "caloriesBurned",
          "duration",
          "date"
        ]
      },
      "UpdateExerciseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Running"
          },
          "caloriesBurned": {
            "type": "number",
            "example": 300
          },
          "duration": {
            "type": "number",
            "example": 30
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15T17:00:00Z"
          }
        }
      },
...
    }
  }
}